requirements.txt
# 核心依赖
streamlit>=1.24.0  # Web应用框架
openai>=1.3.0      # DeepSeek API客户端
markdown>=3.4.0    # Markdown渲染
python-dotenv>=1.0.0  # 环境变量管理
dataclasses>=0.6    # 数据类支持